requirements.txt
Flask
wtforms
passlib
flask_mysqldb
mailchimp3
requests